Fix integer overflow with movie detail fields #2

+++ app/src/main/java/com/example/android/flixtrove/service/model/MovieDetailResponse.java
@@ -19,7 +19,7 @@ public class MovieDetailResponse {
     private Object belongsToCollection;
     @SerializedName("budget")
     @Expose
-    private Integer budget;
+    private Long budget;
     @SerializedName("genres")
     @Expose
     private List<Genre> genres = null;
@@ -58,7 +58,7 @@ public class MovieDetailResponse {
     private String releaseDate;
     @SerializedName("revenue")
     @Expose
-    private Integer revenue;
+    private Long revenue;
     @SerializedName("runtime")
     @Expose
     private Integer runtime;
@@ -108,11 +108,11 @@ public class MovieDetailResponse {
         this.belongsToCollection = belongsToCollection;
     }
 
-    public Integer getBudget() {
+    public Long getBudget() {
         return budget;
     }
 
-    public void setBudget(Integer budget) {
+    public void setBudget(Long budget) {
         this.budget = budget;
     }
 
@@ -212,11 +212,11 @@ public class MovieDetailResponse {
         this.releaseDate = releaseDate;
     }
 
-    public Integer getRevenue() {
+    public Long getRevenue() {
         return revenue;
     }
 
-    public void setRevenue(Integer revenue) {
+    public void setRevenue(Long revenue) {
         this.revenue = revenue;
     }
